UPSTREAM: <carry>: openshift: test/openshift/e2e: don't modify replica count
We were adjusting the replica count when the cluster-autoscaler was
still running which meant that the test would occasionally flake. It
is likely that this bug was introduced in PR #32.

For example, you would occasionally see the following:

```console
I0214 12:51:41.034700 1 scale_up.go:584] Scale-up: setting group size to 3
I0214 12:51:51.129943 1 scale_up.go:584] Scale-up: setting group size to 3
```

Between these two calls we were adjusting the replica count in an
attempt to clean up at the end of the test. But occasionally the
autoscaler would do its scan-of-the-state-of-the-cluster and would up
add new nodes because the replica count was less then desired. When
this condition occurred the node count could never drop below the
initial node count as we just added a further max-min nodes. The test
would eventually timeout trying to assert that the node count matched
the initial node count.

The fix here is to not reset the replica count but instead rely on the
autoscaler to scale down and adjust the replica count naturally; this
change further helps to verify that scale down is working properly.

There are additional smaller fixes here too:

- we set cascading delete in the batch job (i.e., workload)
- we assert that the replica count == the initial replica count
- we explicitly set the clusterautoscaler's ScaleDown config