UPSTREAM: <carry>: openshift: Rework TestControllerNodeGroups
This reworks TestControllerNodeGroups to use subtests. The change is
to ensure we are testing both MachineSets and MachineDeployments in
all cases we construct a nodegroup.

By switching to subtests it is easier to see those places where
machine deployments are not being tested as the the output for each
test function should show that MachineDeployment and MachineSet is
listed in the test name output.

```console
--- PASS: TestControllerNodeGroups (0.61s)
    --- PASS: TestControllerNodeGroups/MachineSet (0.30s)
        --- PASS: TestControllerNodeGroups/MachineSet/errors_with_bad_annotations (0.10s)
        --- PASS: TestControllerNodeGroups/MachineSet/success_with_zero_bounds (0.10s)
        --- PASS: TestControllerNodeGroups/MachineSet/success_with_positive_bounds (0.10s)
    --- PASS: TestControllerNodeGroups/MachineDeployment (0.30s)
        --- PASS: TestControllerNodeGroups/MachineDeployment/errors_with_bad_annotations (0.10s)
        --- PASS: TestControllerNodeGroups/MachineDeployment/success_with_zero_bounds (0.10s)
        --- PASS: TestControllerNodeGroups/MachineDeployment/success_with_positive_bounds (0.10s)
```